add tree config
